+++ CHANGELOG.md
@@ -16,7 +16,10 @@ is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/). Releases a
 versioned in accordance with [Semantic
 Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [0.8.0] "pass the calimari" - Unreleased
+
+### Added
+- `servo.__codename__` constant now contains the release codename.
 
 ### Fixed
 - Progress tracking now handled zero length durations appropriately (e.g., in warmup, settlement, etc).